+++ packages/__vital__/src/styleguide/Page/StyleGuideTemplate/_default.tsx
@@ -15,12 +15,13 @@
 import { vitalTypography } from '@bodiless/vital-elements';
 import { asStyleGuideTemplateToken, vitalStyleGuideTemplate } from '@bodiless/vital-templates';
 import {
-  flowHoc, H2, replaceWith, as, P
+  flowHoc, H2, replaceWith, as, P, Ul,
 } from '@bodiless/fclasses';
 import React from 'react';
 
 const Subtitle = as(vitalTypography.H2, 'pt-8')(H2);
 const Para = as('pt-4')(P);
+const List = as('pt-4')(Ul);
 
 const Examples = () => (
   <>
@@ -31,29 +32,24 @@ const Examples = () => (
       menu on the new page.
     </Para>
     <Subtitle>Global Elements</Subtitle>
-    <Para>
-      <ul>
-        <li><a href="./layout">Layout</a></li>
-        <li><a href="./header">Header</a></li>
-        <li><a href="./footer">Footer</a></li>
-        <li><a href="./editors">Editors</a></li>
-        <li><a href="./editors-monofont">Editors with MonoFont</a></li>
-        <li><a href="./typography">Typography</a></li>
-      </ul>
-    </Para>
+    <List>
+      <li><a href="./layout">Layout</a></li>
+      <li><a href="./header">Header</a></li>
+      <li><a href="./footer">Footer</a></li>
+      <li><a href="./editors">Editors</a></li>
+      <li><a href="./editors-monofont">Editors with MonoFont</a></li>
+      <li><a href="./typography">Typography</a></li>
+    </List>
     <Subtitle>Components</Subtitle>
-    <Para>
-      <ul>
-        <li><a href="./flowcontainer">Flow Container</a></li>
-        <li><a href="./images">Images</a></li>
-      </ul>
-    </Para>
+    <List>
+      <li><a href="./flow-container">Flow Container</a></li>
+      <li><a href="./images">Images</a></li>
+      <li><a href="./list">List</a></li>
+    </List>
     <Subtitle>Templates</Subtitle>
-    <Para>
-      <ul>
-        <li><a href="./generic-template">Generic Template</a></li>
-      </ul>
-    </Para>
+    <List>
+      <li><a href="./generic-template">Generic Template</a></li>
+    </List>
   </>
 );