+++ packages/bodiless-layouts/src/ContentLibrary/ContentLibraryIndicator.tsx
@@ -56,9 +56,9 @@ const LibraryIndicatorBase: FC<LibraryIndicatorProps> = ({ components: C, ...res
 export const LibraryItemIndicatorClean = designable(LibraryItemIndicatorComponents, 'LibraryItemIndicator')(LibraryIndicatorBase);
 
 const asDefaultLibraryItemIndicator = withDesign({
-  Wrapper: addClasses('hidden group-hover:flex bl-text-white bl-bg-gray-900 absolute bl-px-2 bl-py-1 md:-mt-5 md:-ml-5 bl-z-10'),
+  Wrapper: addClasses('bl-hidden group-hover:bl-flex bl-text-white bl-bg-gray-900 bl-absolute bl-px-2 bl-py-1 md:bl-mt-5 md:bl-ml-5 bl-z-10'),
   Icon: flowHoc(
-    addClasses('material-icons bl-mr-2'),
+    addClasses('bl-material-icons bl-mr-2'),
     addProps({ children: 'account_balance' }),
   ),
   Label: addProps({ children: 'Library Item' }),