+++ packages/bodiless-psh/resources/edit/platform.sh
@@ -20,7 +20,7 @@ fi
 
 # Expects the following env variables:
 # APP_VOLUME - the absolute path of the writable volume
-# APP_GIT_REMOTE_URL - the path to the bitbucket git repository
+# APP_GIT_REMOTE_URL - the path to the git repository
 # APP_GIT_USER - the user for git operations
 # APP_GIT_PW - the password for git operations
 # PLATFORM_APP_DIR - the absolute path to the application directory. provided by platform.sh
@@ -29,8 +29,8 @@ fi
 CMD_GIT=/usr/bin/git
 TMP_DIR=${APP_VOLUME}/../tmp
 ROOT_DIR=${APP_VOLUME}/root
-GIT_REMOTE_URL=https://${APP_GIT_USER}:${APP_GIT_PW}@${APP_GIT_REMOTE_URL##https://}
 NPM_CACHE_DIR=${APP_VOLUME}/.npm
+GIT_STORE_CREDENTIAL=${APP_VOLUME}/.credential
 
 invoke () {
   if [[ $(type $1 2>&1) =~ "function" ]]; then
@@ -138,15 +138,16 @@ full_deploy () {
   echo "Full deploy, branch is ${PLATFORM_BRANCH}"
   rm -rf ${ROOT_DIR}
   if [[ ${PLATFORM_BRANCH} =~ ^pr- ]]; then
-    ${CMD_GIT} clone ${GIT_REMOTE_URL} ${ROOT_DIR}
+    ${CMD_GIT} -c credential.helper="!f() { echo username=${APP_GIT_USER}; echo password=${APP_GIT_PW}; }; f" clone ${APP_GIT_REMOTE_URL} ${ROOT_DIR}
     cd ${ROOT_DIR}
     ID=$(echo $PLATFORM_BRANCH | sed s/pr-//g)
     git fetch origin pull/${ID}/head:${PLATFORM_BRANCH}
     git checkout ${PLATFORM_BRANCH}
   else
-    ${CMD_GIT} clone -b ${PLATFORM_BRANCH} ${GIT_REMOTE_URL} ${ROOT_DIR}
+    ${CMD_GIT} -c credential.helper="!f() { echo username=${APP_GIT_USER}; echo password=${APP_GIT_PW}; }; f" clone -b ${PLATFORM_BRANCH} ${APP_GIT_REMOTE_URL} ${ROOT_DIR}
     cd ${ROOT_DIR}
   fi
+  git_store_credential
   ${CMD_GIT} config user.email "${APP_GIT_USER_EMAIL}"
   ${CMD_GIT} config user.name "${APP_GIT_USER}"
 }
@@ -162,9 +163,24 @@ init_npmrc () {
   fi
 }
 
+git_store_credential () {
+  # process credential store only for http based url
+  if  [[ ${APP_GIT_REMOTE_URL} =~ ^http ]] ;
+  then
+    # get host name
+    GIT_HOST=$(echo ${APP_GIT_REMOTE_URL} | awk -F/ '{print $3}')
+    # remove user info, if any
+    GIT_HOST="${GIT_HOST#*:*@}"
+    echo 'https://'${APP_GIT_USER}':'${APP_GIT_PW}'@'${GIT_HOST} > ${GIT_STORE_CREDENTIAL}
+    # set owner permission only
+    chmod 600 ${GIT_STORE_CREDENTIAL}
+    git config --local credential.helper 'store --file='${GIT_STORE_CREDENTIAL}
+  fi
+}
+
 check_branch () {
   if [[ ${PLATFORM_BRANCH} =~ ^pr- ]]; then
-    if [[ ${GIT_REMOTE_URL} =~ github\.com ]]; then
+    if [[ ${APP_GIT_REMOTE_URL} =~ github\.com ]]; then
       return 0
     else
       echo "Edit environments for PR branches are only enabled on GitHub"