+++ packages/bodiless-backend/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.10](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-beta.9...v1.0.0-beta.10) (2022-04-26)
+
+**Note:** Version bump only for package @bodiless/backend
+
+
+
+
+
+# [1.0.0-beta.9](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-beta.5...v1.0.0-beta.9) (2022-04-21)
+
+
+### Features
+
+* Fully disable save by env var. ([#1600](https://github.com/johnsonandjohnson/Bodiless-JS/issues/1600)) ([b19cd73](https://github.com/johnsonandjohnson/Bodiless-JS/commit/b19cd73422bcbb5e82125c4936b7aeb0c7c9cc7e)), closes [#1479](https://github.com/johnsonandjohnson/Bodiless-JS/issues/1479) [#1522](https://github.com/johnsonandjohnson/Bodiless-JS/issues/1522) [#1479](https://github.com/johnsonandjohnson/Bodiless-JS/issues/1479)
+
+
+
+
+
 # [1.0.0-beta.7](https://github.com/johnsonandjohnson/Bodiless-JS/compare/v1.0.0-beta.5...v1.0.0-beta.7) (2022-04-13)
 
 **Note:** Version bump only for package @bodiless/backend